making country context and adding new nav
