+++ components/ui/drawer.tsx
@@ -43,7 +43,7 @@ const DrawerContent = React.forwardRef<
     <DrawerPrimitive.Content
       ref={ref}
       className={cn(
-        "fixed inset-x-0 bottom-0 z-50 mt-24 flex h-auto flex-col rounded-t-[10px] border bg-background",
+        "fixed inset-x-0 bottom-0 z-50 mt-0 flex h-auto flex-col rounded-t-[10px] border bg-background",
         className
       )}
       {...props}
@@ -52,7 +52,7 @@ const DrawerContent = React.forwardRef<
       {children}
     </DrawerPrimitive.Content>
   </DrawerPortal>
-))
+));
 DrawerContent.displayName = "DrawerContent"
 
 const DrawerHeader = ({
